Move live server test into project tests directory.
