+++ channels/test/__init__.py
@@ -1,2 +1,2 @@
-from .base import TransactionChannelTestCase, ChannelTestCase, Client, apply_routes  # NOQA isort:skip
+from .base import TransactionChannelTestCase, ChannelTestCase, ChannelLiveServerTestCase, Client, apply_routes  # NOQA isort:skip
 from .http import HttpClient  # NOQA isort:skip
